% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcDREAM_helperFunctions.R
\name{AdaptpCR}
\alias{AdaptpCR}
\title{Adapts pCR values}
\usage{
AdaptpCR(CR, delta, lCR, settings, Npop)
}
\arguments{
\item{CR}{Vector of crossover probabilities. Needs to be of length nCR.}

\item{delta}{vector with differences}

\item{lCR}{values to weight delta}

\item{settings}{settings list}

\item{Npop}{number of chains.}
}
\value{
Matrix with CR values
}
\description{
Adapts pCR values
}
\keyword{internal}
