% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcFrancesco.R
\name{AM}
\alias{AM}
\title{The Adaptive Metropolis Algorithm}
\usage{
AM(startValue, iterations, nBI, parmin, parmax, FUN, f, eps = 0)
}
\arguments{
\item{startValue}{vector with the start values for the algorithm}

\item{iterations}{iterations to run}

\item{nBI}{number of burnin}

\item{parmin}{minimum values for the parameter vector}

\item{parmax}{maximum values for the parameter vector}

\item{FUN}{function to be sampled from}

\item{f}{scaling factor}

\item{eps}{small number to avoid singularity}
}
\description{
The Adaptive Metropolis Algorithm (Haario et al. 2001)
}
\references{
Haario, Heikki, Eero Saksman, and Johanna Tamminen. "An adaptive Metropolis algorithm." Bernoulli (2001): 223-242.
}

