% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcRun.R
\name{applySettingsDefault}
\alias{applySettingsDefault}
\title{Provides the default settings for the different samplers in runMCMC}
\usage{
applySettingsDefault(settings = NULL, sampler = "DEzs", check = FALSE)
}
\arguments{
\item{settings}{optional list with parameters that will be used instead of the defaults}

\item{sampler}{one of the samplers in \code{\link{runMCMC}}}

\item{check}{logical determines whether parameters should be checked for consistency}
}
\description{
Provides the default settings for the different samplers in runMCMC
}
\details{
see \code{\link{runMCMC}}
}

