% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{beta.search}
\alias{beta.search}
\title{Auxiliary function for adaptive SMC algorithm}
\usage{
beta.search(
  ess,
  target.ess,
  posteriorValues,
  importanceValues,
  oldInter,
  curWeights,
  curExp,
  tol = 1
)
}
\description{
Auxiliary function for adaptive SMC algorithm
}
\keyword{internal}
