% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcTwalk_helperFunctions.R
\name{betaFun}
\alias{betaFun}
\title{Helper function for calculating beta}
\usage{
betaFun(at)
}
\arguments{
\item{at}{"traverse" move proposal parameter.}
}
\description{
Helper function for calculating beta
}
