% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCoda.R
\name{convertCoda}
\alias{convertCoda}
\title{Convert coda::mcmc objects to BayesianTools::mcmcSampler}
\usage{
convertCoda(sampler, names = NULL, info = NULL, likelihood = NULL)
}
\arguments{
\item{sampler}{an object of class mcmc or mcmc.list}

\item{names}{a vector with parameter names (optional)}

\item{info}{a matrix (or list with matrices for mcmc.list objects) with three columns containing log posterior, log likelihood and log prior of the sampler for each time step (optional; but see Details)}

\item{likelihood}{likelihood function used for sampling (see Details)}
}
\description{
Function to support plotting and diagnostic functions for coda::mcmc objects.
}
\details{
The parameter 'likelihood' is optional for most functions but can be needed e.g for  \code{\link{DIC}} function.
Also, the parameter information is typically optional for most uses. However, for certain functions (e.g. \code{\link{MAP}}), the matrix or single columns (e.g. log posterior) are necessary for diagnostics.
}
