% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMcmcSamplerList.R
\name{createMcmcSamplerList}
\alias{createMcmcSamplerList}
\title{Convenience function to create an object of class mcmcSamplerList from a list of mcmc samplers}
\usage{
createMcmcSamplerList(mcmcList)
}
\arguments{
\item{mcmcList}{list of objects, each of which is an mcmcSampler}
}
\value{
object of class "mcmcSamplerList"
}
\description{
Convenience function to create an object of class mcmcSamplerList from a list of mcmc samplers
}
\author{
Florian Hartig
}
