% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSEM.R
\name{createMixWithDefaults}
\alias{createMixWithDefaults}
\title{Allows to mix a given parameter vector with a default parameter vector}
\usage{
createMixWithDefaults(pars, defaults, locations)
}
\arguments{
\item{pars}{vector with new parameter values}

\item{defaults}{vector with defaukt parameter values}

\item{locations}{indices of the new parameter values}
}
\description{
Allows to mix a given parameter vector with a default parameter vector
}

