% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPosterior.R
\name{createPosterior}
\alias{createPosterior}
\title{Creates a standardized posterior class}
\usage{
createPosterior(prior, likelihood)
}
\arguments{
\item{prior}{prior class}

\item{likelihood}{Log likelihood density}
}
\description{
Creates a standardized posterior class
}
\details{
Function is internally used in \code{\link{createBayesianSetup}} to create a standarized posterior class.
}

