% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPrior.R
\name{createPriorDensity}
\alias{createPriorDensity}
\title{Fits a density function to a multivariate sample}
\usage{
createPriorDensity(x, method = "multivariate", eps = 1e-10, lower = NULL,
  upper = NULL, best = NULL)
}
\arguments{
\item{x}{a matrix or BayesianOutput}

\item{method}{method to generate prior - default and currently only option is multivariate}

\item{eps}{numerical}

\item{lower}{vector with lower bounds of parameter}

\item{upper}{vector with upper bounds of parameter}

\item{best}{vector with "best" values of parameter}
}
\description{
Fits a density function to a multivariate sample
}
\examples{
# Create a BayesianSetup
ll <- generateTestDensityMultiNormal(sigma = "no correlation")
bayesianSetup = createBayesianSetup(likelihood = ll, lower = rep(-10, 3), upper = rep(10, 3))

settings = list(iterations = 1000)
out <- runMCMC(bayesianSetup = bayesianSetup, sampler = "Metropolis", settings = settings)

# now use the output as a new prior TODO
#bayesianSetup = createBayesianSetup(likelihood = ll, prior = out)
#out <- runMCMC(bayesianSetup = bayesianSetup, sampler = "Metropolis", settings = settings)

}
\seealso{
\code{\link{createPrior}} \cr
         \code{\link{createBetaPrior}} \cr
         \code{\link{createTruncatedNormalPrior}} \cr
         \code{\link{createUniformPrior}} \cr
         \code{\link{createBayesianSetup}} \cr
}

