% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSMCSamplerList.R
\name{createSmcSamplerList}
\alias{createSmcSamplerList}
\title{Convenience function to create an object of class SMCSamplerList from a list of mcmc samplers}
\usage{
createSmcSamplerList(...)
}
\arguments{
\item{...}{a list of MCMC samplers}
}
\value{
a list of class smcSamplerList with each object being an smcSampler
}
\description{
Convenience function to create an object of class SMCSamplerList from a list of mcmc samplers
}

