% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPrior.R
\name{createUniformPrior}
\alias{createUniformPrior}
\title{Convenience function to create a simple uniform prior distribution}
\usage{
createUniformPrior(lower, upper, best = NULL)
}
\arguments{
\item{lower}{vector of lower prior range for all parameters}

\item{upper}{vector of upper prior range for all parameters}

\item{best}{vector with "best" values for all parameters}
}
\description{
Convenience function to create a simple uniform prior distribution
}
\note{
for details see \code{\link{createPrior}}
}
\examples{
prior <- createUniformPrior(c(0,0),c(0.4,5))
prior$density(c(2,3))
prior$density(c(0.2,2))
}
\seealso{
\code{\link{createPriorDensity}}, \code{\link{createPrior}}, \code{\link{createBetaPrior}}, \code{\link{createTruncatedNormalPrior}}, \code{\link{createBayesianSetup}}
}

