% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcFrancesco.R
\name{DR}
\alias{DR}
\title{The Delayed Rejection Algorithm}
\usage{
DR(startValue, iterations, nBI, parmin, parmax, f1, f2, FUN)
}
\arguments{
\item{startValue}{vector with the start values for the algorithm}

\item{iterations}{iterations to run}

\item{nBI}{number of burnin}

\item{parmin}{minimum values for the parameter vector}

\item{parmax}{maximum values for the parameter vector}

\item{f1}{scaling factor for first proposal}

\item{f2}{scaling factor for second proposal}

\item{FUN}{function to be sampled from}
}
\description{
The Delayed Rejection Algorithm (Tierney and Mira, 1999)
}
\references{
Tierney, Luke, and Antonietta Mira. "Some adaptive Monte Carlo methods for Bayesian inference." Statistics in medicine 18.1718 (1999): 2507-2515.
}

