% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposalGenerator.R
\name{factorMatrice}
\alias{factorMatrice}
\title{factorMatrice}
\usage{
factorMatrice(sigma, method)
}
\arguments{
\item{sigma}{sigma}

\item{method}{either "eigen", "svd" or "chol"}
}
\description{
factorMatrice
}
\keyword{internal}
