% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcDREAM_helperFunctions.R
\name{generateCRvalues}
\alias{generateCRvalues}
\title{Generates matrix of CR values based on pCR}
\usage{
generateCRvalues(pCR, settings, Npop)
}
\arguments{
\item{pCR}{Vector of crossover probabilities. Needs to be of length nCR.}

\item{settings}{settings list}

\item{Npop}{number of chains}
}
\value{
Matrix with CR values
}
\description{
Generates matrix of CR values based on pCR
}
\keyword{internal}
