% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockUpdate.R
\name{getBlock}
\alias{getBlock}
\title{Determine the parameters in the block update}
\usage{
getBlock(blockSettings)
}
\arguments{
\item{blockSettings}{settings for block update}
}
\value{
vector containing the parameter to be updated
}
\description{
Determine the parameters in the block update
}
