% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockUpdate.R
\name{getBlockSettings}
\alias{getBlockSettings}
\title{getblockSettings}
\usage{
getBlockSettings(blockUpdate)
}
\arguments{
\item{blockUpdate}{input settings}
}
\value{
list with block settings
}
\description{
Transforms the original settings to settings used in the model runs
}
\keyword{internal}
