% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcConfidence.R
\name{getCredibleIntervals}
\alias{getCredibleIntervals}
\title{Calculate confidence region from an MCMC or similar sample}
\usage{
getCredibleIntervals(sampleMatrix, quantiles = c(0.025, 0.975))
}
\arguments{
\item{sampleMatrix}{matrix of outcomes. Could be parameters or predictions}

\item{quantiles}{quantiles to be calculated}
}
\description{
Calculate confidence region from an MCMC or similar sample
}
\seealso{
\code{\link{getPredictiveDistribution}} \cr
         \code{\link{getPredictiveIntervals}} \cr
}
\author{
Florian Hartig
}
