% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{getDharmaResiduals}
\alias{getDharmaResiduals}
\title{Creates a DHARMa object}
\usage{
getDharmaResiduals(model, parMatrix, numSamples, observed, error, plot = TRUE)
}
\arguments{
\item{model}{function that calculates model predictions for a given parameter vector}

\item{parMatrix}{a parameter matrix from which the simulations will be generated}

\item{numSamples}{the number of samples}

\item{observed}{a vector of observed values}

\item{error}{function with signature f(mean, par) that generates error expectations from mean model predictions. Par is a vector from the matrix with the parameter samples (full length). f needs to know which of these parameters are parameters of the error function}

\item{plot}{logical, determining whether the simulated residuals should be plotted}
}
\description{
Creates a DHARMa object
}
\author{
Tankred Ott
}
