% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcRun.R
\name{getMetropolisDefaultSettings}
\alias{getMetropolisDefaultSettings}
\title{Returns Metropolis default settings}
\usage{
getMetropolisDefaultSettings()
}
\description{
Returns Metropolis default settings
}
\author{
Tankred Ott
}
\keyword{internal}
