% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPanels}
\alias{getPanels}
\title{getPanels}
\usage{
getPanels(x)
}
\arguments{
\item{x}{the desired number of panels}
}
\description{
Calculates the argument x for par(mfrow = x) for a desired number of panels
}
\author{
Florian Hartig
}
