% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcConfidence.R
\name{getPredictiveIntervals}
\alias{getPredictiveIntervals}
\title{Calculates Bayesian credible (confidence) and predictive intervals based on parameter sample}
\usage{
getPredictiveIntervals(parMatrix, model, thin = 1000, quantiles = c(0.025,
  0.975), error = NULL)
}
\arguments{
\item{parMatrix}{matrix of parameter values}

\item{model}{model / function to calculate predictions. Outcome should be a vector}

\item{thin}{thinning}

\item{quantiles}{quantiles to calculate}

\item{error}{function that accepts a vector as in observation. If supplied, will calculate also predictive intervals additional to credible intervals}
}
\description{
Calculates Bayesian credible (confidence) and predictive intervals based on parameter sample
}
\seealso{
\code{\link{getPredictiveDistribution}} \cr
         \code{\link{getCredibleIntervals}} \cr
}
\author{
Florian Hartig
}
