% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposalGenerator.R
\name{getRmvnorm}
\alias{getRmvnorm}
\title{Produce multivariate normal proposal}
\usage{
getRmvnorm(n = 1, R)
}
\arguments{
\item{n}{n}

\item{R}{R}
}
\value{
X
}
\description{
Produce multivariate normal proposal
}
