% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSetup}
\alias{getSetup}
\title{Function to get the setup from a bayesianOutput}
\usage{
getSetup(x)
}
\arguments{
\item{x}{bayesianOutput}
}
\value{
bayesianSetup
}
\description{
Function to get the setup from a bayesianOutput
}
\author{
Tankred Ott
}
\keyword{internal}
