% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcTwalk_helperFunctions.R
\name{Gfun}
\alias{Gfun}
\title{Helper function for blow and hop moves}
\usage{
Gfun(case, npSel, pSel, h, x, x2)
}
\arguments{
\item{case}{Type of Twalk move. Either "hop" or "blow"}

\item{npSel}{number of parameters that are changed.}

\item{pSel}{vector containing information about which parameters are changed.}

\item{h}{Parameter for "blow" and hop moves}

\item{x}{parameter vector of chain 1}

\item{x2}{parameter vector of chain 2}
}
\description{
Helper function for blow and hop moves
}
\references{
Christen, J. Andres, and Colin Fox. "A general purpose sampling algorithm for continuous distributions (the t-walk)." Bayesian Analysis 5.2 (2010): 263-281.
}
\keyword{internal}
