% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classLikelihood.R
\name{likelihoodAR1}
\alias{likelihoodAR1}
\title{AR1 type likelihood function}
\usage{
likelihoodAR1(predicted, observed, sd, a)
}
\arguments{
\item{predicted}{vector of predicted values}

\item{observed}{vector of observed values}

\item{sd}{standard deviation of the iid normal likelihood}

\item{a}{temporal correlation in the AR1 model}
}
\description{
AR1 type likelihood function
}
\note{
The AR1 model considers the process y(t) = a y(t-1) + E, e = iid N(0,sd), |a| < 1. At the moment, no NA are allowed in the time series
}
\author{
Florian Hartig
}

