% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classLikelihood.R
\name{likelihoodIidNormal}
\alias{likelihoodIidNormal}
\title{Normal / Gaussian Likelihood function}
\usage{
likelihoodIidNormal(predicted, observed, sd)
}
\arguments{
\item{predicted}{vector of predicted values}

\item{observed}{vector of observed values}

\item{sd}{standard deviation of the iid normal likelihood}
}
\description{
Normal / Gaussian Likelihood function
}
\author{
Florian Hartig
}

