% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcFrancesco.R
\name{M}
\alias{M}
\title{The Metropolis Algorithm}
\usage{
M(startValue, iterations, nBI, parmin, parmax, f, FUN, consoleUpdates = 1000)
}
\arguments{
\item{startValue}{vector with the start values for the algorithm}

\item{iterations}{iterations to run}

\item{nBI}{number of burnin}

\item{parmin}{minimum values for the parameter vector}

\item{parmax}{maximum values for the parameter vector}

\item{f}{scaling factor}

\item{FUN}{function to be sampled from}

\item{consoleUpdates}{interger, determines the frequency with which sampler progress is printed to the console}
}
\description{
The Metropolis Algorithm (Metropolis et al. 1953)
}
\references{
Metropolis, Nicholas, et al. "Equation of state calculations by fast computing machines." The journal of chemical physics 21.6 (1953): 1087-1092.
}

