% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP.R
\name{MAP}
\alias{MAP}
\title{calculates the Maxiumum APosteriori value (MAP)}
\usage{
MAP(bayesianOutput, ...)
}
\arguments{
\item{bayesianOutput}{an object of class BayesianOutput (mcmcSampler, smcSampler, or mcmcList)}

\item{...}{optional values to be passed on the the getSample function}
}
\description{
calculates the Maxiumum APosteriori value (MAP)
}
\details{
Currently, this function simply returns the parameter combination with the highest posterior in the chain. A more refined option would be to take the MCMC sample and do additional calculations, e.g. use an optimizer, a kernel density estimator, or some other tool to search / interpolate around the best value in the chain.
}
\seealso{
\code{\link{WAIC}}, \code{\link{DIC}}, \code{\link{marginalLikelihood}}
}
\author{
Florian Hartig
}
