% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarginals.R, R/plotMarginals_2.R
\name{marginalPlotDensity}
\alias{marginalPlotDensity}
\alias{marginalPlotDensity}
\title{Plot marginals as densities}
\usage{
marginalPlotDensity(posteriorMat, priorMat = NULL, xrange = NULL,
  col = c("#FF5000A0", "#4682B4A0"), singlePanel = TRUE, ...)

marginalPlotDensity(posteriorMat, priorMat = NULL, xrange = NULL,
  col = c("#FF5000A0", "#4682B4A0"), singlePanel = TRUE, ...)
}
\arguments{
\item{posteriorMat}{matrix with samples as rows and parameters as columns}

\item{priorMat}{matrix (optional) with samples as rows and parameters as columns}

\item{xrange}{vector or matrix (optional), determining the plotting range, with parameters as columns and min, max as rows}

\item{col}{vector of colors for posterior and}

\item{singlePanel}{logical, determining whether the parameter should be plotted in a single panel or each in its own panel}

\item{posteriorMat}{matrix with samples as rows and parameters as columns}

\item{priorMat}{matrix (optional) with samples as rows and parameters as columns}

\item{xrange}{vector or matrix (optional), determining the plotting range, with parameters as columns and min, max as rows}

\item{col}{vector of colors for posterior and}

\item{singlePanel}{logical, determining whether the parameter should be plotted in a single panel or each in its own panel}
}
\description{
Plot marginals as densities

Plot marginals as densities
}
\author{
Tankred Ott

Tankred Ott
}
\keyword{internal}
