% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarginals.R
\name{marginalPlotViolin}
\alias{marginalPlotViolin}
\title{Plot marginals as violin plot}
\usage{
marginalPlotViolin(posteriorMat, priorMat = NULL, xrange = NULL,
  col = c("#FC8D62FF", "#A6D85499"), singlePanel = TRUE, ...)
}
\arguments{
\item{posteriorMat}{matrix with samples as rows and parameters as columns}

\item{priorMat}{matrix (optional) with samples as rows and parameters as columns}

\item{xrange}{vector or matrix (optional), determining the plotting range, with parameters as columns and min, max as rows}

\item{col}{vector of colors for posterior and}

\item{singlePanel}{logical, determining whether the parameter should be plotted in a single panel or each in its own panel}
}
\description{
Plot marginals as violin plot
}
\author{
Tankred Ott
}
\keyword{internal}
