% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcMultipleChains.R
\name{mcmcMultipleChains}
\alias{mcmcMultipleChains}
\title{Run multiple chains}
\usage{
mcmcMultipleChains(bayesianSetup, settings, sampler)
}
\arguments{
\item{bayesianSetup}{object of class "BayesianSetup"}

\item{settings}{list containing settings for sampler}

\item{sampler}{character, either "Metropolis" or "DE"}
}
\value{
list containing the single runs ($sampler) and the chains in a coda::mcmc.list ($mcmc.list)
}
\description{
Run multiple chains
}
\keyword{internal}
