% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmcOutput}
\alias{mcmcOutput}
\title{Example MCMC output}
\format{An object of class \code{mcmcSampler} (inherits from \code{bayesianOutput}) of length 7.}
\usage{
mcmcOutput
}
\description{
Example MCMC output
}
\keyword{datasets}
