% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMcmcSampler.R
\name{summary.mcmcSampler}
\alias{summary.mcmcSampler}
\alias{print.mcmcSampler}
\alias{plot.mcmcSampler}
\title{Summmary of MCMC output}
\usage{
\method{summary}{mcmcSampler}(object, ...)

\method{print}{mcmcSampler}(x, ...)

\method{plot}{mcmcSampler}(x, ...)
}
\arguments{
\item{...}{not implemented}
}
\description{
Creates a summary table of a MCMC output

Plots MCMC output
}
\author{
Stefan Paul

Florian Hartig
}
