% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classBayesianOutput.R
\name{mergeChains}
\alias{mergeChains}
\title{Merge Chains}
\usage{
mergeChains(l, ...)
}
\arguments{
\item{l}{a list with objects that can be accessed with \code{\link{getSample}}}

\item{...}{arguments to be passed on to \code{\link{getSample}}}
}
\value{
a matrix
}
\description{
Merge a list of outputs from MCMC / SMC samplers
}
\details{
The function merges a list of outputs from MCMC / SMC samplers into a single matrix. Requirement is that the list contains classes for which the \code{\link{getSample}} function works
}
\author{
Florian Hartig
}
