% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{mutate}
\alias{mutate}
\title{Auxiliary function for particle mutation}
\usage{
mutate(
  setup,
  particles,
  proposalGenerator,
  posteriorValues,
  importanceDensity,
  method,
  steps,
  proposalScale,
  adaptive = TRUE,
  b = 1e-04,
  min.pars = NULL,
  max.pars = NULL,
  Z = NULL
)
}
\description{
Auxiliary function for particle mutation
}
\keyword{internal}
