% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMcmcSampler.R, R/classMcmcSamplerList.R
\name{plot.mcmcSampler}
\alias{plot.mcmcSampler}
\alias{plot.mcmcSamplerList}
\title{Plots of MCMC output}
\usage{
\method{plot}{mcmcSampler}(x, ...)

\method{plot}{mcmcSamplerList}(x, ...)
}
\arguments{
\item{x}{object of class mcmcSampler or mcmcSamplerList}

\item{...}{additional options passed to tracePlot}
}
\description{
Plots MCMC output
}
\section{Functions}{
\itemize{
\item \code{plot(mcmcSamplerList)}: 

}}
\seealso{
\code{\link{getSample.mcmcSampler}}
}
\author{
Florian Hartig
}
