% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSMCSamplerList.R, R/classSmcSampler.R
\name{plot.smcSamplerList}
\alias{plot.smcSamplerList}
\alias{plot.smcSampler}
\title{Plots of smcSampler output}
\usage{
\method{plot}{smcSamplerList}(x, ...)

\method{plot}{smcSampler}(x, ...)
}
\arguments{
\item{x}{object of class mcmcSampler or mcmcSamplerList}

\item{...}{additional options passed to tracePlot}
}
\description{
Plots smcSampler output
}
\section{Functions}{
\itemize{
\item \code{plot(smcSamplerList)}: 

}}
