% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarginals_old.R
\name{plotHist}
\alias{plotHist}
\title{plot histogram}
\usage{
plotHist(x, at = NULL, col = "orangered", border = "black")
}
\arguments{
\item{x}{matrix (e.g. constructed with \code{\link{createBreakMat}}) or dataframe containing the lower bounds, the upper bounds, and the frequencies of the breaks in the columns and the individual breaks in the rows.}

\item{at}{y position of the histogram. If NULL a new plot will be generated.}

\item{col}{color of the histogram}

\item{border}{border color}
}
\description{
A simple custom histogram plotting function
}
\author{
Tankred Ott
}
\keyword{internal}
