% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plots a time series, with the option to include confidence and prediction band}
\usage{
plotTimeSeries(observed = NULL, predicted = NULL, x = NULL,
  confidenceBand = NULL, predictionBand = NULL, xlab = "Time",
  ylab = "Observed / predicted values", ...)
}
\arguments{
\item{observed}{observed values}

\item{predicted}{predicted values}

\item{x}{optional values for x axis (time)}

\item{confidenceBand}{matrix with confidenceBand}

\item{predictionBand}{matrix with predictionBand}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{...}{further arguments passed to \code{\link[graphics]{plot}}}
}
\description{
Plots a time series, with the option to include confidence and prediction band
}
\author{
Florian Hartig
}
\seealso{
\code{\link{plotTimeSeriesResults}}  \cr
         \code{\link{marginalPlot}} \cr
         \code{\link{tracePlot}} \cr
         \code{\link{correlationPlot}}
}

