% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeriesResiduals}
\alias{plotTimeSeriesResiduals}
\title{Plots residuals of a time series}
\usage{
plotTimeSeriesResiduals(residuals, x = NULL, main = "residuals")
}
\arguments{
\item{residuals}{x}

\item{x}{optional values for x axis (time)}

\item{main}{title of the plot}
}
\description{
Plots residuals of a time series
}
\author{
Florian Hartig
}
