% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classBayesianSetup.R
\name{print.BayesianSetup}
\alias{print.BayesianSetup}
\title{Print an object of BayesianSetup}
\usage{
\method{print}{BayesianSetup}(x, ...)
}
\arguments{
\item{x}{an object of class BayesianSetup}

\item{...}{not implemented}
}
\description{
print function for objects of class BayesianSetup
}
\seealso{
\code{\link{checkBayesianSetup}}
}
\author{
Maximilian Pichler
}
