% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMcmcSampler.R
\name{print.mcmcSampler}
\alias{print.mcmcSampler}
\title{Print of MCMC output}
\usage{
\method{print}{mcmcSampler}(x, ...)
}
\description{
Print MCMC output
}
\seealso{
\code{\link{getSample.mcmcSampler}}
}
\author{
Florian Hartig
}
