% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPrior.R
\name{print.prior}
\alias{print.prior}
\title{Print object of prior class}
\usage{
\method{print}{prior}(x, ...)
}
\arguments{
\item{x}{object of a prior}

\item{...}{additional parameters}
}
\description{
print function for objects of class prior
}
\seealso{
\code{\link{createPrior}}
}
\author{
Maximilian Pichler
}
