% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSmcSampler.R
\name{print.smcSampler}
\alias{print.smcSampler}
\title{Print of smcSampler output}
\usage{
\method{print}{smcSampler}(x, ...)
}
\description{
Print smcSampler output
}
\seealso{
\code{\link{getSample.smcSampler}}
}
\author{
Florian Hartig
}
