% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcTwalk_helperFunctions.R
\name{propFun}
\alias{propFun}
\title{Helper function to create proposal}
\usage{
propFun(case, Npar, pn1, x, x2, beta = NULL, aw = NULL)
}
\arguments{
\item{case}{Type of Twalk move. Either "walk", "traverse", "hop" or "blow"}

\item{Npar}{number of parameters}

\item{pn1}{Probability determining the number of parameters that are changed.}

\item{x}{parameter vector of chain 1}

\item{x2}{parameter vector of chain 2}

\item{beta}{parameter for "traverse" move proposals.}

\item{aw}{"walk" move proposal parameter.}
}
\description{
Helper function to create proposal
}
\references{
Christen, J. Andres, and Colin Fox. "A general purpose sampling algorithm for continuous distributions (the t-walk)." Bayesian Analysis 5.2 (2010): 263-281.
}

