% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{resample}
\alias{resample}
\title{Auxiliary function for adaptive SMC algorithm}
\usage{
resample(weights, method = "multinomial")
}
\description{
Auxiliary function for adaptive SMC algorithm
}
\keyword{internal}
