% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale}
\alias{rescale}
\title{Rescale}
\usage{
rescale(x, from, to)
}
\arguments{
\item{x}{Vector of values}

\item{from}{vector, interval of which x are elements. from\link{1} must be the lower, from\link{2} the upper bound.}

\item{to}{vector, interval to which the elements should be scaled. to\link{1} must be the lower, to\link{2} the upper bound.}
}
\description{
Rescales values in the interval "from" (lower, upper) to the new interval "to" (lower, upper).
}
\author{
Tankred Ott
}
\keyword{internal}
