% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classBayesianSetup.R
\name{rmBayesianSetup}
\alias{rmBayesianSetup}
\title{Function to remove objects of class BayesianSetup}
\usage{
rmBayesianSetup(bayesianSetup, envir = NULL)
}
\arguments{
\item{bayesianSetup}{object of class BayesianSetup}

\item{envir}{environment}
}
\description{
Function removes the BayesianSetup and closes 
the parallel executer (if available)
}

