% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcMetropolis.R
\name{sampleMetropolis}
\alias{sampleMetropolis}
\title{gets samples while adopting the MCMC proposal generator}
\usage{
sampleMetropolis(mcmcSampler, iterations)
}
\arguments{
\item{mcmcSampler}{an mcmcSampler}

\item{iterations}{iterations}
}
\description{
Function to sample with cobinations of the basic Metropolis-Hastings MCMC algorithm (Metropolis et al., 1953), a variation of the adaptive Metropolis MCMC (Haario et al., 2001), the delayed rejection algorithm (Tierney & Mira, 1999), and the delayed rejection adaptive Metropolis algorithm (DRAM, Haario et al), and the Metropolis within Gibbs
}
\author{
Florian Hartig
}
\keyword{internal}
