% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaleMatrix}
\alias{scaleMatrix}
\title{Function to scale matrices}
\usage{
scaleMatrix(mat, min, max)
}
\arguments{
\item{mat}{matrix to scale}

\item{min}{minimum value}

\item{max}{maximum value}
}
\value{
sclaed matrix
}
\description{
Function to scale matrices
}
\keyword{internal}
