% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcRun.R
\name{setupStartProposal}
\alias{setupStartProposal}
\title{Help function to find starvalues and proposalGenerator settings}
\usage{
setupStartProposal(proposalGenerator = NULL, bayesianSetup, settings)
}
\arguments{
\item{proposalGenerator}{proposal generator}

\item{bayesianSetup}{either an object of class bayesianSetup created by \code{\link{createBayesianSetup}} (recommended), or a log target function}

\item{settings}{list with settings}
}
\description{
Help function to find starvalues and proposalGenerator settings
}

