% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classBayesianSetup.R
\name{stopParallel}
\alias{stopParallel}
\title{Function to close cluster in BayesianSetup}
\usage{
stopParallel(bayesianSetup)
}
\arguments{
\item{bayesianSetup}{object of class BayesianSetup}
}
\description{
Function closes 
the parallel executer (if available)
}

