% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMcmcSampler.R, R/classMcmcSamplerList.R,
%   R/classSmcSampler.R
\name{summary.mcmcSampler}
\alias{summary.mcmcSampler}
\alias{summary.mcmcSamplerList}
\alias{summary.smcSampler}
\title{Summmary of MCMC output}
\usage{
\method{summary}{mcmcSampler}(object, ...)

\method{summary}{mcmcSamplerList}(object, ...)

\method{summary}{smcSampler}(object, ...)
}
\arguments{
\item{...}{not implemented}
}
\description{
Creates a summary table of a MCMC output

Creates a summary table of a 'smcSampler' output
}
\section{Functions}{
\itemize{
\item \code{summary(mcmcSamplerList)}: Method for mcmcSamplerList

\item \code{summary(smcSampler)}: Summary for smcSampler objects

}}
\seealso{
\code{\link{getSample.mcmcSampler}}
}
\author{
Stefan Paul

Florian Hartig
}
