% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSMCSamplerList.R, R/classSmcSampler.R
\name{summary.smcSamplerList}
\alias{summary.smcSamplerList}
\alias{summary.smcSampler}
\title{Summary for class 'smcSampler'}
\usage{
\method{summary}{smcSamplerList}(object, ...)

\method{summary}{smcSampler}(object, ...)
}
\arguments{
\item{object}{object of class smcSampler or smcSamplerList}

\item{...}{additional parameters}
}
\description{
Creates a summary table of a 'smcSampler' output
}
\section{Functions}{
\itemize{
\item \code{summary(smcSamplerList)}: 

}}
\author{
Florian Hartig
}
