% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcTwalk_helperFunctions.R
\name{sumSquare}
\alias{sumSquare}
\title{Helper function for sum of x*x}
\usage{
sumSquare(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
Helper function for sum of x*x
}
\keyword{internal}
