% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testDensityGelmanMeng}
\alias{testDensityGelmanMeng}
\title{Normal likelihood}
\usage{
testDensityGelmanMeng(x, A = 1, B = 0, C1 = 3, C2 = 3, log = TRUE)
}
\description{
A non-elliptical, bivariate density function proposed by Gelman and Meng (1991).
}
