% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testDensityInfinity}
\alias{testDensityInfinity}
\title{Test function infinity ragged}
\usage{
testDensityInfinity(x, error = F)
}
\arguments{
\item{x}{2-dim parameter vector}

\item{error}{should error or infinity be returned}
}
\description{
Test function infinity ragged
}
\seealso{
\code{\link{generateTestDensityMultiNormal}} \cr
         \code{\link{testDensityBanana}}
}
\author{
Florian Hartig
}
