% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testDensityMultiNormal}
\alias{testDensityMultiNormal}
\title{3d Mutivariate Normal likelihood}
\usage{
testDensityMultiNormal(x, sigma = "strongcorrelation")
}
\arguments{
\item{x}{a parameter vector of arbitrary length}

\item{sigma}{either a correlation matrix, or "strongcorrelation", or "no correlation"}
}
\description{
3d Mutivariate Normal likelihood
}
