% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testDensityNormal}
\alias{testDensityNormal}
\title{Normal likelihood}
\usage{
testDensityNormal(x, sum = T)
}
\arguments{
\item{x}{a parameter vector of arbitrary length}

\item{sum}{if likelihood should be summed or not}
}
\description{
Normal likelihood
}
\author{
Florian Hartig
}
