% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testLinearModel}
\alias{testLinearModel}
\title{Fake model, returns a ax + b linear response to 2-param vector}
\usage{
testLinearModel(x, env = NULL)
}
\arguments{
\item{x}{2-dim parameter vector}

\item{env}{optional, environmental covariate}
}
\description{
Fake model, returns a ax + b linear response to 2-param vector
}
\examples{
x = c(1,2)
y = testLinearModel(x)
plot(y)
}
\seealso{
\code{\link{generateTestDensityMultiNormal}} \cr
         \code{\link{testDensityBanana}}
}
\author{
Florian Hartig
}
