% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thinMatrix}
\alias{thinMatrix}
\title{Function to thin matrices}
\usage{
thinMatrix(mat, thin = "auto")
}
\arguments{
\item{mat}{matrix to thin}

\item{thin}{thinning parameter}
}
\value{
thinned matrix
}
\description{
Function to thin matrices
}

