% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostic.R
\name{plotDiagnostic}
\alias{plotDiagnostic}
\title{Diagnostic Plot}
\usage{
plotDiagnostic(out, lty = 1, lwd = 1, numSamples = 10)
}
\arguments{
\item{out}{object of class "bayesianOutput"  An object of class BayesianSetup}

\item{lty}{line typ}

\item{lwd}{line width}

\item{numSamples}{for calculating WAIC, default = 10 because of high computational cost}
}
\section{Functions}{
\itemize{
\item \code{plotDiagnostic}: function plots the DIC, WAIC, mPSRF and the trace of the parameters in dependence of Iterations. DIC, WAIC are plotted separately for the chains and the trace plots also for the internal chains.
}}

\note{
This function can also be used for sensitivity analysis of an arbitrary output - just create a BayesianSetup with this output.
}
\examples{
# Create bayesian setup with  
bayesianSetup <- createBayesianSetup(likelihood = testDensityNormal, 
                                     prior = createUniformPrior(lower = rep(-10,2),
                                                                upper = rep(10,2)))

# running MCMC

out = runMCMC(bayesianSetup = bayesianSetup)

# diagnostic plots

plotDiagnostic(out)

}
\author{
Maximilian Pichler
}
