% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrace.R
\name{tracePlot}
\alias{tracePlot}
\title{Trace plot for MCMC class}
\usage{
tracePlot(sampler, thin = "auto", ...)
}
\arguments{
\item{sampler}{an object of class MCMC sampler}

\item{thin}{determines the thinning intervall of the chain}

\item{...}{additional parameters to pass on to the \code{\link{getSample}}, for example parametersOnly =F, or start = 1000}
}
\description{
Trace plot for MCMC class
}
\seealso{
\code{\link{marginalPlot}} \cr
         \code{\link{plotTimeSeries}} \cr
         \code{\link{correlationPlot}}
}

