% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcTwalk_helperFunctions.R
\name{Twalksteps}
\alias{Twalksteps}
\title{Main function that is executing and evaluating the moves}
\usage{
Twalksteps(case, Npar, FUN, x, Eval, x2, Eval2, at, aw, pn1)
}
\arguments{
\item{case}{Type of Twalk move. Either "walk", "traverse", "hop" or "blow"}

\item{Npar}{number of parameters}

\item{FUN}{Log posterior density}

\item{x}{parameter vector of chain 1}

\item{Eval}{last evaluation of x}

\item{x2}{parameter vector of chain 2}

\item{Eval2}{last evaluation of x}

\item{at}{"traverse" move proposal parameter.}

\item{aw}{"walk" move proposal parameter.}

\item{pn1}{Probability determining the number of parameters that are changed.}
}
\description{
Main function that is executing and evaluating the moves
}
\references{
Christen, J. Andres, and Colin Fox. "A general purpose sampling algorithm for continuous distributions (the t-walk)." Bayesian Analysis 5.2 (2010): 263-281.
}
\keyword{internal}
