% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockUpdate.R
\name{updateGroups}
\alias{updateGroups}
\title{Determine the groups of correlated parameters}
\usage{
updateGroups(chain, blockSettings)
}
\arguments{
\item{chain}{MCMC chain including only the parameters (not logP,ll, logP)}

\item{blockSettings}{a list with settings}
}
\value{
groups
}
\description{
Determine the groups of correlated parameters
}
\author{
Stefan Paul
}
\keyword{internal}
