% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposalGenerator.R
\name{updateProposalGenerator}
\alias{updateProposalGenerator}
\title{To update settings of an existing proposal genenerator}
\usage{
updateProposalGenerator(
  proposal,
  chain = NULL,
  message = F,
  eps = 1e-10,
  manualScaleAdjustment = 1
)
}
\arguments{
\item{proposal}{an object of class proposalGenerator}

\item{chain}{a chain to create the covariance matrix from (optional)}

\item{message}{whether to print an updating message}

\item{eps}{numeric tolerance for covariance}

\item{manualScaleAdjustment}{optional adjustment for the covariance scale (multiplicative)}
}
\description{
To update settings of an existing proposal genenerator
}
\details{
The this function can be applied in 2 ways 1) update the covariance given an MCMC chain, and 2) update the proposal generator after parameters have been changed
}
