% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vsemC}
\alias{vsemC}
\title{C version of the VSEM model}
\usage{
vsemC(par, PAR)
}
\arguments{
\item{par}{parameter vector}

\item{PAR}{Photosynthetically active radiation (PAR) MJ /m2 /day}
}
\description{
C version of the VSEM model
}
