% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSEM.R
\name{VSEMcreateLikelihood}
\alias{VSEMcreateLikelihood}
\title{Create an example dataset, and from that a likelihood or posterior for the VSEM model}
\usage{
VSEMcreateLikelihood(likelihoodOnly = F, plot = F, selection = c(1:6, 12))
}
\arguments{
\item{likelihoodOnly}{logical, decides whether to create only a likelihood, or a full bayesianSetup with uniform priors.}

\item{plot}{logical, decides whether data should be plotted}

\item{selection}{vector containing the indices of the selected parameters}
}
\description{
Create an example dataset, and from that a likelihood or posterior for the VSEM model
}
\details{
The purpose of this function is to be able to conveniently create a likelihood for the VSEM model for demonstration purposes. The function creates example data --> likelihood --> BayesianSetup, where the latter is the
}
\author{
Florian Hartig
}
