% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSEM.R
\name{VSEMcreatePAR}
\alias{VSEMcreatePAR}
\title{Create a random radiation (PAR) time series}
\usage{
VSEMcreatePAR(days = 1:(3 * 365))
}
\arguments{
\item{days}{days to calculate the PAR for}
}
\description{
Create a random radiation (PAR) time series
}
\author{
David Cameron, R implementation by Florian Hartig
}
