% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correctThin}
\alias{correctThin}
\title{Checks if thin is conistent with nTotalSamples samples and if not corrects it.}
\usage{
correctThin(nTotalSamples, thin, fraction = 5000)
}
\arguments{
\item{nTotalSamples}{total number of rows/samples}

\item{thin}{thinning}
}
\description{
Checks if thin is conistent with nTotalSamples samples and if not corrects it.
}
\details{
Checks if the thin argument is consistent with the data consisting of nTotalSamples samples/rows and corrects thin if not.
}
\author{
Tankred Ott

Tankred Ott
}
