% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPrior.R
\name{createBetaPrior}
\alias{createBetaPrior}
\title{Convenience function to create a beta prior}
\usage{
createBetaPrior(a, b, lower = 0, upper = 1)
}
\arguments{
\item{a}{shape1 of the beta distribution}

\item{b}{shape2 of the beta distribution}

\item{lower}{lower values for the parameters}

\item{upper}{upper values for the parameters}
}
\description{
Convenience function to create a beta prior
}
\details{
This creates a beta prior, assuming that lower / upper values for parameters are are fixed. The beta is the calculated relative to this lower / upper space.
}
\note{
for details see \code{\link{createPrior}}
}
\seealso{
\code{\link{createPriorDensity}} \cr
         \code{\link{createPrior}} \cr
         \code{\link{createTruncatedNormalPrior}} \cr
         \code{\link{createUniformPrior}} \cr
         \code{\link{createBayesianSetup}} \cr
}

