% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPrior.R
\name{createPrior}
\alias{createPrior}
\title{Creates a standardized prior class}
\usage{
createPrior(density = NULL, sampler = NULL, lower = NULL, upper = NULL,
  best = NULL)
}
\arguments{
\item{density}{Prior density}

\item{sampler}{Sampling function for density (optional)}

\item{lower}{vector with lower bounds of parameters}

\item{upper}{vector with upper bounds of parameter}

\item{best}{vector with "best" parameter values}
}
\description{
Creates a standardized prior class
}
\note{
min and max don't update the absolute values of the prior density. Hence, if absolute values are a concern, trucated distributions should be used for the density
}
\examples{
# Create a general prior distribution by specifying an arbitrary density function and a
# corresponding sampling function
density = function(par){
  d1 = dunif(par[1], -2,6, log =TRUE)
  d2 = dnorm(par[2], mean= 2, sd = 3, log =TRUE)
  return(d1 + d2)
}

# The sampling is optional but recommended because the MCMCs can generate automatic starting
# conditions if this is provided
sampler = function(n=1){
  d1 = runif(n, -2,6)
  d2 = rnorm(n, mean= 2, sd = 3)
  return(cbind(d1,d2))
}

prior <- createPrior(density = density, sampler = sampler, 
                     lower = NULL, upper = NULL, best = NULL)

}
\seealso{
\code{\link{createPriorDensity}} \cr
         \code{\link{createBetaPrior}} \cr
         \code{\link{createUniformPrior}} \cr
         \code{\link{createTruncatedNormalPrior}}\cr
         \code{\link{createBayesianSetup}}\cr
}

