% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPanels}
\alias{getPanels}
\title{Calculates the panel combination for par(mfrow = )}
\usage{
getPanels(x)
}
\arguments{
\item{x}{the desired number of panels}
}
\description{
Calculates the panel combination for par(mfrow = )
}
\author{
Florian Hartig
}
