% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcConfidence.R
\name{getPredictiveDistribution}
\alias{getPredictiveDistribution}
\title{Calculates predictive distribution based on the parameters}
\usage{
getPredictiveDistribution(parMatrix, model, thin = 1000)
}
\arguments{
\item{parMatrix}{matrix of parameter values}

\item{model}{model / function to calculate predictions. Outcome should be a vector}

\item{thin}{thinning}
}
\description{
Calculates predictive distribution based on the parameters
}
\seealso{
\code{\link{getPredictiveIntervals}} \cr
         \code{\link{getCredibleIntervals}} \cr
}
\author{
Florian Hartig
}
