% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarginals.R
\name{histMarginal}
\alias{histMarginal}
\title{histogram for marginalPlot}
\usage{
histMarginal(posterior, prior = NULL, breaks = 15, at = 1, col = NULL)
}
\arguments{
\item{posterior}{vector of posterior values}

\item{prior}{vector of prior values or NULL}

\item{breaks}{integer, determining the number of breaks}

\item{at}{y position of the histograms}

\item{col}{vector determining posterior and prior color}
}
\description{
Internal function to plot histograms for the marginalPlot function
}
\author{
Tankred Ott
}
