% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-Numerics.R
\name{logSumExp}
\alias{logSumExp}
\title{Funktion to compute log(sum(exp(x))}
\usage{
logSumExp(x, mean = F)
}
\arguments{
\item{x}{values}

\item{mean}{logical, determines whether the mean should be used}
}
\description{
Funktion to compute log(sum(exp(x))
}
\author{
Florian Hartig
}
