% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcFrancesco.R
\name{M}
\alias{M}
\title{The Metropolis Algorithm}
\usage{
M(startValue = NULL, iterations = 10000, nBI = 0, parmin = NULL,
  parmax = NULL, f = 1, FUN, consoleUpdates = 1000)
}
\arguments{
\item{startValue}{vector with the start values for the algorithm. Can be NULL if FUN is of class BayesianSetup. In this case startValues are sampled from the prior.}

\item{iterations}{iterations to run}

\item{nBI}{number of burnin}

\item{parmin}{minimum values for the parameter vector or NULL if FUN is of class BayesianSetup}

\item{parmax}{maximum values for the parameter vector or NULL if FUN is of class BayesianSetup}

\item{f}{scaling factor}

\item{FUN}{function to be sampled from or object of class bayesianSetup}

\item{consoleUpdates}{interger, determines the frequency with which sampler progress is printed to the console}
}
\description{
The Metropolis Algorithm (Metropolis et al. 1953)
}
\references{
Metropolis, Nicholas, et al. "Equation of state calculations by fast computing machines." The journal of chemical physics 21.6 (1953): 1087-1092.
}
\author{
Francesco Minunno
}
\keyword{internal}
