% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codaFunctions.R
\name{makeObjectClassCodaMCMC}
\alias{makeObjectClassCodaMCMC}
\title{Helper function to change an object to a coda mcmc class,}
\usage{
makeObjectClassCodaMCMC(chain, start = 1, end = numeric(0), thin = 1)
}
\arguments{
\item{chain}{mcmc Chain}

\item{start}{For MCMC samplers, the initial value in the chain. For SMC samplers, initial particle}

\item{end}{For MCMC samplers, the end value in the chain. For SMC samplers, end particle.}

\item{thin}{thinning parameter}
}
\value{
object an object of class coda::mcmc
}
\description{
Helper function to change an object to a coda mcmc class,
}
\details{
Very similar to coda::mcmc but with less overhead
}
\keyword{internal}
