% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classBayesianOutput.R
\name{mergeChains}
\alias{mergeChains}
\title{Merge Chains}
\usage{
mergeChains(l, ...)
}
\arguments{
\item{l}{the list with MCMC outputs}

\item{...}{arguments to be passed on to getSample}
}
\value{
a matrix
}
\description{
Merge a list of MCMCs or chains
}
\details{
The function merges a list of MCMC objects. Requirement is that the list contains classes for which the getSample function works
}
\author{
Florian Hartig
}
