% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metropolisRatio}
\alias{metropolisRatio}
\title{Function to calculate the metropolis ratio}
\usage{
metropolisRatio(LP2, LP1, tempering = 1)
}
\arguments{
\item{LP2}{log posterior old position}

\item{LP1}{log posterior of proposal}

\item{tempering}{value for tempering}
}
\description{
Function to calculate the metropolis ratio
}
\author{
Florian Hartig
}
\keyword{internal}
