% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMcmcSampler.R
\name{plot.mcmcSampler}
\alias{plot.mcmcSampler}
\title{Plots of MCMC output}
\usage{
\method{plot}{mcmcSampler}(x, ...)
}
\arguments{
\item{x}{description}

\item{...}{additional options passed to tracePlot}
}
\description{
Plots MCMC output
}
\seealso{
\code{\link{getSample.mcmcSampler}}
}
\author{
Florian Hartig
}
