% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMcmcSampler.R
\name{print.mcmcSampler}
\alias{print.mcmcSampler}
\title{Prints MCMC output}
\usage{
\method{print}{mcmcSampler}(x, ...)
}
\arguments{
\item{x}{object of class mcmcSampler or mcmcSamplerList}

\item{...}{additional options}
}
\description{
Prints MCMC output
}
\seealso{
\code{\link{getSample.mcmcSampler}}
}
