% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMcmcSampler.R, R/classMcmcSamplerList.R
\name{print.mcmcSampler}
\alias{print.mcmcSampler}
\alias{print.mcmcSamplerList}
\title{Prints MCMC output}
\usage{
\method{print}{mcmcSampler}(x, ...)

\method{print}{mcmcSamplerList}(x, ...)
}
\arguments{
\item{x}{object of class mcmcSampler or mcmcSamplerList}

\item{...}{additional options}
}
\description{
Prints MCMC output
}
\section{Functions}{
\itemize{
\item \code{print(mcmcSamplerList)}: 

}}
\seealso{
\code{\link{getSample.mcmcSampler}}
}
\author{
Florian Hartig
}
