% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSMCSamplerList.R, R/classSmcSampler.R
\name{print.smcSamplerList}
\alias{print.smcSamplerList}
\alias{print.smcSampler}
\title{Print of smcSampler output}
\usage{
\method{print}{smcSamplerList}(x, ...)

\method{print}{smcSampler}(x, ...)
}
\arguments{
\item{x}{object of class mcmcSampler or mcmcSamplerList}

\item{...}{additional options}
}
\description{
Print smcSampler output
}
\section{Functions}{
\itemize{
\item \code{print(smcSamplerList)}: 

}}
\author{
Florian Hartig
}
