% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale}
\alias{rescale}
\title{Rescale}
\usage{
rescale(x, from, to)
}
\arguments{
\item{x}{vector of values to be scaled}

\item{from}{vector of length 2, original interval (lower, upper)}

\item{to}{vector of length 2, target interval (lower, upper)}
}
\description{
Rescales values in the interval "from" (lower, upper) to the new interval "to" (lower, upper).
}
\author{
Tankred Ott
}
\keyword{internal}
