% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarginals.R
\name{rescale}
\alias{rescale}
\title{Rescale}
\usage{
rescale(x, from, to)
}
\arguments{
\item{x}{Vector of values}

\item{from}{vector, interval of which x are elements. from[1] must be the lower, from[2] the upper bound.}

\item{to}{vector, interval to which the elements should be scaled. to[1] must be the lower, to[2] the upper bound.}
}
\author{
Tankred Ott
}
