% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleEquallySpaced}
\alias{sampleEquallySpaced}
\title{Gets n equally spaced samples (rows) from a matrix or vector}
\usage{
sampleEquallySpaced(x, numSamples)
}
\arguments{
\item{x}{matrix or vector}

\item{numSamples}{number of samples (rows) to be drawn}
}
\description{
Gets n equally spaced samples (rows) from a matrix or vector
}
\details{
Gets n equally spaced samples (rows) from a matrix and returns a new matrix (or vector) containing those samples
}
\author{
Tankred Ott
}
\keyword{internal}
