% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMcmcSampler.R
\name{summary.mcmcSampler}
\alias{summary.mcmcSampler}
\title{Summmary of MCMC output}
\usage{
\method{summary}{mcmcSampler}(object, ...)
}
\arguments{
\item{...}{not implemented}
}
\description{
Creates a summary table of a MCMC output
}
\seealso{
\code{\link{getSample.mcmcSampler}}
}
\author{
Stefan Paul
}
