% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMcmcSampler.R, R/classMcmcSamplerList.R
\name{summary.mcmcSampler}
\alias{summary.mcmcSampler}
\alias{summary.mcmcSamplerList}
\title{Summmary of MCMC output}
\usage{
\method{summary}{mcmcSampler}(object, ...)

\method{summary}{mcmcSamplerList}(object, ...)
}
\arguments{
\item{object}{object of class mcmcSampler or mcmcSamplerList}

\item{...}{not implemented}
}
\description{
Creates a summary table of a MCMC output
}
\section{Functions}{
\itemize{
\item \code{summary(mcmcSamplerList)}: 

}}
\seealso{
\code{\link{getSample.mcmcSampler}}
}
\author{
Stefan Paul
}
