% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSmcSampler.R
\name{summary.smcSampler}
\alias{summary.smcSampler}
\title{Summary for class 'smcSampler'}
\usage{
\method{summary}{smcSampler}(object, ...)
}
\description{
Creates a summary table of a 'smcSampler' output
}
\seealso{
\link{\code{getSample.smcSampler}}
}
\author{
Florian Hartig
}
