% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{testDensityBanana}
\alias{testDensityBanana}
\title{Banana-shaped density function}
\usage{
testDensityBanana(p)
}
\arguments{
\item{p}{2-dim parameter vector}
}
\description{
Banana-shaped density function
}
\note{
inspired from package FMEmcmc, sees to go back to Laine M (2008). Adaptive MCMC Methods with Applications in Environmental and Models. Finnish Meteorological Institute Contributions 69. ISBN 978-951-697-662-7.
examples()
}
\author{
Florian Hartig
}
\seealso{
\code{\link{generateTestDensityMultiNormal}} \cr
         \code{\link{testLinearModel}}
}

